AG-17809 Rename the quadrant preset away from "scatter"
Public types drop the Scatter prefix (AgScatterQuadrantPreset ->
AgQuadrantPreset and friends), scatterQuadrantOptions.ts folds into
quadrantOptions.ts, and the enterprise preset directory, module, factory
and options defs follow. The internal presetType string becomes
'quadrant'.

Contained only in nightly betas so far, so no deprecated aliases.
