+++ packages/ag-charts-enterprise/src/preset/quadrant/__image_snapshots__/quadrant-preset-test-ts-quadrant-preset-for-pivot-numeric-it-should-render-to-canvas-as-expected-1-snap.png
