+++ packages/ag-charts-enterprise/src/preset/quadrant/quadrantPresetModule.ts
@@ -3,8 +3,8 @@ import type { PresetModuleDefinition } from 'ag-charts-core';
 import type { AgQuadrantChartOptions, ExtensibleSeriesTheme } from 'ag-charts-types';
 
 import { backgroundRegionsTheme } from '../../features/background-regions/backgroundRegionsTheme';
-import { createScatterQuadrant } from './scatterQuadrantPreset';
-import { scatterQuadrantOptionsDefs } from './scatterQuadrantPresetOptionsDefs';
+import { createQuadrant } from './quadrantPreset';
+import { quadrantOptionsDefs } from './quadrantPresetOptionsDefs';
 
 const sharedThemeTemplate: ExtensibleSeriesTheme<'bubble' | 'scatter'> = {
     axes: {
@@ -42,16 +42,16 @@ const sharedThemeTemplate: ExtensibleSeriesTheme<'bubble' | 'scatter'> = {
     },
 };
 
-export const ScatterQuadrantPresetModule: PresetModuleDefinition<AgQuadrantChartOptions> = {
+export const QuadrantPresetModule: PresetModuleDefinition<AgQuadrantChartOptions> = {
     type: 'preset',
-    name: 'scatter-quadrant',
+    name: 'quadrant',
     enterprise: true,
     dependencies: [],
     version: VERSION,
 
-    options: scatterQuadrantOptionsDefs,
+    options: quadrantOptionsDefs,
 
-    create: createScatterQuadrant,
+    create: createQuadrant,
 
     themeTemplate: {
         scatter: sharedThemeTemplate,
