+++ packages/ag-charts-types/src/main.ts
@@ -49,7 +49,6 @@ export * from './presets/gauge/gaugeOptions';
 export * from './presets/gauge/linearGaugeOptions';
 export * from './presets/gauge/radialGaugeOptions';
 export * from './presets/quadrant/quadrantOptions';
-export * from './presets/quadrant/scatterQuadrantOptions';
 export * from './presets/sparkline/sparklineAxisOptions';
 export * from './presets/sparkline/sparklineOptions';
 export * from './series/cartesian/areaOptions';
